Move functions out of service_constants.h
