+++ browser/ui/browser_commands.cc
@@ -76,6 +76,7 @@
 
 #if BUILDFLAG(IS_WIN)
 #include "brave/components/brave_vpn/common/wireguard/win/service_constants.h"
+#include "brave/components/brave_vpn/common/wireguard/win/service_details.h"
 #include "brave/components/brave_vpn/common/wireguard/win/storage_utils.h"
 #endif  // BUILDFLAG(ENABLE_BRAVE_VPN)
