+++ components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_connection_api.cc
@@ -10,7 +10,7 @@
 
 #include "brave/components/brave_vpn/browser/connection/wireguard/brave_vpn_wireguard_connection_api_base.h"
 #include "brave/components/brave_vpn/common/win/utils.h"
-#include "brave/components/brave_vpn/common/wireguard/win/service_constants.h"
+#include "brave/components/brave_vpn/common/wireguard/win/service_details.h"
 #include "brave/components/brave_vpn/common/wireguard/win/wireguard_utils.h"
 
 namespace brave_vpn {
