+++ components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_service/service/wireguard_service_runner.cc
@@ -26,7 +26,7 @@
 #include "base/threading/thread_restrictions.h"
 #include "base/win/scoped_com_initializer.h"
 #include "brave/components/brave_vpn/browser/connection/wireguard/win/brave_vpn_wireguard_service/service/brave_wireguard_manager.h"
-#include "brave/components/brave_vpn/common/wireguard/win/service_constants.h"
+#include "brave/components/brave_vpn/common/wireguard/win/service_details.h"
 #include "brave/components/brave_vpn/common/wireguard/win/wireguard_utils.h"
 
 namespace brave_vpn {
