+++ components/brave_vpn/common/wireguard/win/BUILD.gn
@@ -18,8 +18,9 @@ midl("brave_wireguard_manager_idl") {
 
 source_set("win") {
   sources = [
-    "service_constants.cc",
     "service_constants.h",
+    "service_details.cc",
+    "service_details.h",
     "storage_utils.cc",
     "storage_utils.h",
     "wireguard_utils.cc",
