+++ components/brave_vpn/common/wireguard/win/service_details.h
@@ -1,0 +1,26 @@
+/* Copyright (c) 2023 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#ifndef BRAVE_COMPONENTS_BRAVE_VPN_COMMON_WIREGUARD_WIN_SERVICE_DETAILS_H_
+#define BRAVE_COMPONENTS_BRAVE_VPN_COMMON_WIREGUARD_WIN_SERVICE_DETAILS_H_
+
+#include <guiddef.h>
+#include <string>
+
+#include "base/files/file_path.h"
+#include "base/version.h"
+
+namespace brave_vpn {
+const CLSID& GetBraveVpnWireguardServiceClsid();
+const IID& GetBraveVpnWireguardServiceIid();
+std::wstring GetBraveVpnWireguardTunnelServiceName();
+std::wstring GetBraveVpnWireguardServiceName();
+std::wstring GetBraveVpnWireguardServiceDisplayName();
+base::FilePath GetBraveVPNWireguardServiceInstallationPath(
+    const base::FilePath& target_path,
+    const base::Version& version);
+}  // namespace brave_vpn
+
+#endif  // BRAVE_COMPONENTS_BRAVE_VPN_COMMON_WIREGUARD_WIN_SERVICE_DETAILS_H_
